pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.1",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.20.0",
    "uvicorn>=0.40.0",
]
